requirements.txt
click==8.1.3
matplotlib==3.7.1
numpy==1.24.2
python-dotenv==1.0.0
scikit_learn==1.2.2
setuptools==67.6.1
torch==2.1.0.dev20230324
torchvision==0.16.0.dev20230324
